refactor move paypal standard settings
