+++ includes/admin/settings/class-settings-gateways.php
@@ -35,8 +35,8 @@ if ( ! class_exists( 'Give_Settings_Gateways' ) ) :
 
 			// Do not use main form for this tab.
 			if ( give_get_current_setting_tab() === $this->id ) {
-				add_action( 'give_admin_field_gateway_notice', array( $this, 'render_gateway_notice' ), 10, 2 );
-				add_action( 'give_admin_field_enabled_gateways', array( $this, 'render_enabled_gateways' ), 10, 2 );
+				add_action( 'give_admin_field_gateway_notice', [ $this, 'render_gateway_notice' ], 10, 2 );
+				add_action( 'give_admin_field_enabled_gateways', [ $this, 'render_enabled_gateways' ], 10, 2 );
 			}
 		}
 
@@ -47,153 +47,80 @@ if ( ! class_exists( 'Give_Settings_Gateways' ) ) :
 		 * @return array
 		 */
 		public function get_settings() {
-			$settings        = array();
+			$settings        = [];
 			$current_section = give_get_current_setting_section();
 
 			switch ( $current_section ) {
-				case 'paypal-standard':
-					$settings = array(
-						// Section 2: PayPal Standard.
-						array(
-							'type' => 'title',
-							'id'   => 'give_title_gateway_settings_2',
-						),
-						array(
-							'name' => __( 'PayPal Email', 'give' ),
-							'desc' => __( 'Enter the email address associated with your PayPal account to connect with the gateway.', 'give' ),
-							'id'   => 'paypal_email',
-							'type' => 'email',
-						),
-						array(
-							'name' => __( 'PayPal Page Style', 'give' ),
-							'desc' => __( 'Enter the name of the PayPal page style to use, or leave blank to use the default.', 'give' ),
-							'id'   => 'paypal_page_style',
-							'type' => 'text',
-						),
-						array(
-							'name'    => __( 'PayPal Transaction Type', 'give' ),
-							'desc'    => __( 'Nonprofits must verify their status to withdraw donations they receive via PayPal. PayPal users that are not verified nonprofits must demonstrate how their donations will be used, once they raise more than $10,000. By default, GiveWP transactions are sent to PayPal as donations. You may change the transaction type using this option if you feel you may not meet PayPal\'s donation requirements.', 'give' ),
-							'id'      => 'paypal_button_type',
-							'type'    => 'radio_inline',
-							'options' => array(
-								'donation' => __( 'Donation', 'give' ),
-								'standard' => __( 'Standard Transaction', 'give' ),
-							),
-							'default' => 'donation',
-						),
-						array(
-							'name'    => __( 'Billing Details', 'give' ),
-							'desc'    => __( 'If enabled, required billing address fields are added to PayPal Standard forms. These fields are not required by PayPal to process the transaction, but you may have a need to collect the data. Billing address details are added to both the donation and donor record in GiveWP.', 'give' ),
-							'id'      => 'paypal_standard_billing_details',
-							'type'    => 'radio_inline',
-							'default' => 'disabled',
-							'options' => array(
-								'enabled'  => __( 'Enabled', 'give' ),
-								'disabled' => __( 'Disabled', 'give' ),
-							),
-						),
-						array(
-							'name'    => __( 'PayPal IPN Verification', 'give' ),
-							'desc'    => __( 'If enabled, IPN (Instant Payment Notification) messages sent to your site from PayPal are verified with an extra (background) step. The IPN is what marks PayPal donations as complete on GiveWP\'s side. If donations are not getting marked as complete, disabling this extra verification step can resolve it. Only disable this setting to resolve the pending donation issue, since it is technically less secure.', 'give' ),
-							'id'      => 'paypal_verification',
-							'type'    => 'radio_inline',
-							'default' => 'enabled',
-							'options' => array(
-								'enabled'  => __( 'Enabled', 'give' ),
-								'disabled' => __( 'Disabled', 'give' ),
-							),
-						),
-						array(
-							'id'      => 'paypal_invoice_prefix',
-							'name'    => esc_html__( 'Invoice ID Prefix', 'give' ),
-							'desc'    => esc_html__( 'Please enter a prefix for your invoice numbers. If you use your PayPal account for multiple fundraising platforms or ecommerce stores, ensure this prefix is unique. PayPal will not allow orders or donations with the same invoice number.', 'give' ),
-							'type'    => 'text',
-							'default' => 'GIVE-',
-						),
-						array(
-							'name'  => __( 'PayPal Standard Gateway Settings Docs Link', 'give' ),
-							'id'    => 'paypal_standard_gateway_settings_docs_link',
-							'url'   => esc_url( 'http://docs.givewp.com/settings-gateway-paypal-standard' ),
-							'title' => __( 'PayPal Standard Gateway Settings', 'give' ),
-							'type'  => 'give_docs_link',
-						),
-						array(
-							'type' => 'sectionend',
-							'id'   => 'give_title_gateway_settings_2',
-						),
-					);
-					break;
-
 				case 'offline-donations':
-					$settings = array(
+					$settings = [
 						// Section 3: Offline gateway.
-						array(
+						[
 							'type' => 'title',
 							'id'   => 'give_title_gateway_settings_3',
-						),
-						array(
+						],
+						[
 							'name'    => __( 'Collect Billing Details', 'give' ),
 							'desc'    => __( 'If enabled, required billing address fields are added to Offline Donation forms. These fields are not required to process the transaction, but you may have a need to collect the data. Billing address details are added to both the donation and donor record in GiveWP. ', 'give' ),
 							'id'      => 'give_offline_donation_enable_billing_fields',
 							'type'    => 'radio_inline',
 							'default' => 'disabled',
-							'options' => array(
+							'options' => [
 								'enabled'  => __( 'Enabled', 'give' ),
 								'disabled' => __( 'Disabled', 'give' ),
-							),
-						),
-						array(
+							],
+						],
+						[
 							'name'    => __( 'Offline Donation Instructions', 'give' ),
 							'desc'    => __( 'The Offline Donation Instructions are a chance for you to educate the donor on how to best submit offline donations. These instructions appear directly on the form, and after submission of the form. Note: You may also customize the instructions on individual forms as needed.', 'give' ),
 							'id'      => 'global_offline_donation_content',
 							'default' => give_get_default_offline_donation_content(),
 							'type'    => 'wysiwyg',
-							'options' => array(
+							'options' => [
 								'textarea_rows' => 6,
-							),
-						),
-						array(
+							],
+						],
+						[
 							'name'  => esc_html__( 'Offline Donations Settings Docs Link', 'give' ),
 							'id'    => 'offline_gateway_settings_docs_link',
 							'url'   => esc_url( 'http://docs.givewp.com/offlinegateway' ),
 							'title' => __( 'Offline Gateway Settings', 'give' ),
 							'type'  => 'give_docs_link',
-						),
-						array(
+						],
+						[
 							'type' => 'sectionend',
 							'id'   => 'give_title_gateway_settings_3',
-						),
-					);
+						],
+					];
 					break;
 
 				case 'gateways-settings':
-					$settings = array(
+					$settings = [
 						// Section 1: Gateways.
-						array(
+						[
 							'id'   => 'give_title_gateway_settings_1',
 							'type' => 'title',
-						),
-						array(
+						],
+						[
 							'id'   => 'gateway_notice',
 							'type' => 'gateway_notice',
-						),
-						array(
+						],
+						[
 							'name'    => __( 'Test Mode', 'give' ),
 							'desc'    => __( 'If enabled, donations are processed through the sandbox/test accounts configured in each gateway\'s settings. This prevents having to use real money for tests. See the payment gateway documentation for instructions on configuring sandbox accounts.', 'give' ),
 							'id'      => 'test_mode',
 							'type'    => 'radio_inline',
 							'default' => 'disabled',
-							'options' => array(
+							'options' => [
 								'enabled'  => __( 'Enabled', 'give' ),
 								'disabled' => __( 'Disabled', 'give' ),
-							),
-						),
-						array(
+							],
+						],
+						[
 							'name' => __( 'Enabled Gateways', 'give' ),
 							'desc' => __( 'Enable your payment gateway. Can be ordered by dragging.', 'give' ),
 							'id'   => 'gateways',
 							'type' => 'enabled_gateways',
-						),
+						],
 
 						/**
 						 * "Enabled Gateways" setting field contains gateways label setting but when you save gateway settings then label will not save
@@ -201,12 +128,12 @@ if ( ! class_exists( 'Give_Settings_Gateways' ) ) :
 						 *
 						 * This setting will not render on admin setting screen but help internal code to recognize "gateways_label"  setting and add them to give setting when save.
 						 */
-						array(
+						[
 							'name' => __( 'Gateways Label', 'give' ),
 							'desc' => '',
 							'id'   => 'gateways_label',
 							'type' => 'gateways_label_hidden',
-						),
+						],
 
 						/**
 						 * "Enabled Gateways" setting field contains default gateway setting but when you save gateway settings then this setting will not save
@@ -214,25 +141,25 @@ if ( ! class_exists( 'Give_Settings_Gateways' ) ) :
 						 *
 						 * This setting will not render on admin setting screen but help internal code to recognize "default_gateway"  setting and add them to give setting when save.
 						 */
-						array(
+						[
 							'name' => __( 'Default Gateway', 'give' ),
 							'desc' => __( 'The gateway that will be selected by default.', 'give' ),
 							'id'   => 'default_gateway',
 							'type' => 'default_gateway_hidden',
-						),
+						],
 
-						array(
+						[
 							'name'  => __( 'Gateways Docs Link', 'give' ),
 							'id'    => 'gateway_settings_docs_link',
 							'url'   => esc_url( 'http://docs.givewp.com/settings-gateways' ),
 							'title' => __( 'Gateway Settings', 'give' ),
 							'type'  => 'give_docs_link',
-						),
-						array(
+						],
+						[
 							'id'   => 'give_title_gateway_settings_1',
 							'type' => 'sectionend',
-						),
-					);
+						],
+					];
 					break;
 			}
 
@@ -262,11 +189,10 @@ if ( ! class_exists( 'Give_Settings_Gateways' ) ) :
 		 * @return array
 		 */
 		public function get_sections() {
-			$sections = array(
+			$sections = [
 				'gateways-settings' => __( 'Gateways', 'give' ),
-				'paypal-standard'   => __( 'PayPal Standard', 'give' ),
 				'offline-donations' => __( 'Offline Donations', 'give' ),
-			);
+			];
 
 			return apply_filters( 'give_get_sections_' . $this->id, $sections );
 		}
@@ -348,7 +274,7 @@ if ( ! class_exists( 'Give_Settings_Gateways' ) ) :
 		public function render_enabled_gateways( $field, $settings ) {
 			$id              = $field['id'];
 			$gateways        = give_get_ordered_payment_gateways( give_get_payment_gateways() );
-			$gateways_label  = give_get_option( 'gateways_label', array() );
+			$gateways_label  = give_get_option( 'gateways_label', [] );
 			$default_gateway = give_get_option( 'default_gateway', current( array_keys( $gateways ) ) );
 
 			ob_start();
